% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{alive}
\alias{alive}
\title{Teams Still Alive}
\usage{
alive(tournament, teams = 1L:n_teams(tournament))
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}

\item{teams}{A numerical vector of teams numbers.}
}
\value{
A subset of \code{teams} containing only teams who have not yet lost.
}
\description{
List teams that have not yet lost.
}
