% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{all_games}
\alias{all_games}
\title{Produce a table of games and scores}
\usage{
all_games(tournament, game_scores, keep.all = FALSE)
}
\arguments{
\item{game_scores}{A data frame containing games scores. The following columns must
exist: \code{home} (chr), \code{away} (chr), \code{hscore} (int), \code{ascore} (int).}

\item{keep.all}{A logical indicating whether rows should be kept for games
for which the opponents are not yet determined.}
}
\value{
A data frame with columns \code{game_number}, \code{home}, \code{away}, \code{ascore}, \code{hscore},
\code{description}, plus any additional columns that were present in \code{game_scores}.
}
\description{
Produce a table of games and scores
}
