% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{build_entry_matrix}
\alias{build_entry_matrix}
\title{Create Entry Matrix}
\usage{
build_entry_matrix(E, ext = c("M", "W"))
}
\arguments{
\item{E}{a list of lists. The nested lists should have slots for
"name", "email", "dept", "time", "time", "points", "teamsLogical", and "teamsLogicalW".
The latter two are logical vectors with as many values as teams in the tournament}

\item{ext}{Either "M" or "W" indicating which tournament.}
}
\description{
Convert a list of lists into a matrix representing each entrants selections.
Attributes are used to store some additional information.
}
\examples{
EM <- build_entry_matrix(Ent, "M")
EW <- build_entry_matrix(Ent, "W")
}
