% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{competitors}
\alias{competitors}
\title{List of entries with teams}
\usage{
competitors(entries, division = c("M", "W"), by = c("email", "name"))
}
\description{
Some entrants don't select teams in all brackets. This determines list of true
competetitors.
}
