% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{contest_standings}
\alias{contest_standings}
\title{Tabular report of contest standings
Tabular report of contest standings}
\usage{
contest_standings(tournament, entries, bracket)
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}

\item{entries}{a matrix with a row for each entrant and a column for each team.
Entries indicate if the team was chosen by the entrant.}
}
\description{
Tabular report of contest standings
Tabular report of contest standings
}
