% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{dust}
\alias{dust}
\title{Compute tie-breaking dust}
\usage{
dust(x)
}
\arguments{
\item{x}{a numeric vector of numbers between 0 and 6 (inclusive)}
}
\value{
a vector the same length of \code{x} with a bit of "dust" added to the values.
}
\description{
Add a small amount to the number of wins so that summing the dusted wins
will correctly sort participants with tie-breaking rules accounted for.
The total amount of dust in any round is less than the dust for one team
in the next round.
}
