% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{home_team_name}
\alias{home_team_name}
\alias{away_team_name}
\alias{team_names}
\title{Extract team names with various properties}
\usage{
home_team_name(tournament, game = 1L:n_games(tournament))

away_team_name(tournament, game = 1L:n_games(tournament))

team_names(tournament)
}
\arguments{
\item{tournament}{A tournament object}

\item{game}{A numeric vector of game numbers}
}
\value{
For \code{home_team_name()} and \code{away_team_name()},
a character vector of home/away team names for the specified games.

For \code{team_names()} a character vector of team names specified by the numbers
in `teams.
}
\description{
Extract team names with various properties
}
