% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{initialize_tournament}
\alias{initialize_tournament}
\title{Create a tournament with no games played}
\usage{
initialize_tournament(names, seeds = NULL, label = "G")
}
\arguments{
\item{rounds}{Number of rounds in a complete single elimination tournament}
}
\description{
Create a tournament with no games played
}
