% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_entries_from_files}
\alias{load_entries_from_files}
\title{Load Entries From Files}
\usage{
load_entries_from_files(
  tournament,
  path,
  year = 2022,
  pattern = paste0("Entry-", year, ".*rds"),
  keep.all = FALSE
)
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}

\item{path}{directory in which entry files are located.}

\item{year}{used to create the default value of pattern.}

\item{pattern}{regex pattern.}
}
\description{
Load entries from files in a directory that match a specified pattern.
}
