% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_entries_from_pins}
\alias{load_entries_from_pins}
\title{Load Entries From Pins}
\usage{
load_entries_from_pins(
  board,
  year = 2023,
  pattern = paste0("NCAA-", year, "-entry-"),
  keep.all = FALSE
)
}
\arguments{
\item{board}{a pins board}

\item{year}{used to create the default value of pattern.}

\item{pattern}{regex pattern.}
}
\description{
Load entries from pins that match a specified pattern.
}
