% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_game_scores}
\alias{load_game_scores}
\title{Load Games Scores}
\usage{
load_game_scores(files)
}
\arguments{
\item{files}{A vector of csv files names. Each file should have coluns \code{game_number} (int),
\code{winner_01} (0 if "home" wins, 1 if "away" wins),
\code{home} (chr), \code{away} (chr), \code{hscore} (int), and \code{ascore (int)}}
}
\value{
a data frame containing information about each game. \code{winner} and \code{loser} columns are computed from
the scores and \code{home} and \code{away}.
}
\description{
Load game scores from a CSV file or files.
}
