% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{load_scores_file}
\alias{load_scores_file}
\title{Load Scores}
\usage{
load_scores_file(file)
}
\arguments{
\item{file}{The file name of a csv with columms \code{game_number}, \code{winner_01}, (0 if "home" wins, 1 if "away" wins),
\code{home}, \code{away}, \code{hscore}, and \code{ascore}}
}
\value{
a data frame containing information about each game. \code{winner} and \code{loser} columns are computed from
the scores and \code{home} and \code{away}.
}
\description{
Load game scores from a CSV file.
}
