% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{round_by_round}
\alias{round_by_round}
\title{Round-by-round scores}
\usage{
round_by_round(tournament, entries)
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}

\item{entries}{a matrix with a row for each entrant and a column for each team.
Entries indicate if the team was chosen by the entrant.}
}
\value{
A matrix with a row for each entrant and a column for each round of the tournament.
Entries indicate the number of wins for each player in each of the rounds.
}
\description{
Compute an entrant-by-round matrix of round scores for each entrant.
}
