% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_matrix}
\alias{scenario_matrix}
\title{Create Scenario Matrix}
\usage{
scenario_matrix(n_games = 15, safe = TRUE)
}
\arguments{
\item{n_games}{number of games with undetermined outcome}

\item{safe}{a logical indicating whether use should be blocked when \code{n_games} is large.}
}
\description{
Create a bit matrix of possible outcomes of unknown games.
Note: This gets large fast, so the input is capped at 15
}
\examples{
scenario_matrix(5)
}
