% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{scores}
\alias{scores}
\title{Compute Scores}
\usage{
scores(tournament, entries, dust = TRUE)
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}

\item{entries}{a matrix with a row for each entrant and a column for each team.
Entries indicate if the team was chosen by the entrant.}

\item{dust}{a logical indicating whether tie breaking dust should be added to the scores.}
}
\description{
Compute scores for each entrant. Each entrant receives one point for each win by one of the
entrants selected teams.
}
