% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{scores_table}
\alias{scores_table}
\title{Create scores table for a modeling competition}
\usage{
scores_table(tournament, entries)
}
\description{
Create scores table for a modeling competition.
}
