% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\docType{data}
\name{seedOrder}
\alias{seedOrder}
\alias{seedCostOld}
\alias{seedCost}
\title{Seed order and costs}
\format{
An object of class \code{numeric} of length 16.

An object of class \code{numeric} of length 16.

An object of class \code{numeric} of length 16.
}
\usage{
seedOrder

seedCostOld

seedCost
}
\description{
Numeric vectors giving seed order and costs for the NCAA modeling competition.
}
\keyword{datasets}
