% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{team_names}
\alias{team_names}
\title{Team Names}
\usage{
team_names(tournament)
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}
}
\description{
Get team names from a tournament object.
}
