% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{tournament_init}
\alias{tournament_init}
\title{Create a tournament with no games played}
\usage{
tournament_init(names, seeds = NULL, label = "G")
}
\arguments{
\item{rounds}{Number of rounds in a complete single elimination tournament}
}
\description{
Create a tournament with no games played
}
