% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{update_tournament}
\alias{update_tournament}
\title{Update tournament based on game outcomes}
\usage{
update_tournament(tournament, games, results)
}
\arguments{
\item{tournament}{a 0-1 vector of tournament game outcomes. Games are played from high index to
low index. Game g pairs the winners of games \code{2*g} and \code{2*g + 1}.}

\item{games}{a vector of game indices}

\item{results}{a 0-1 vector of results for \code{games}. 0 indicates that the lower indexed team won.
1 indicates that the higher indexed team won.}
}
\description{
Update tournament based on game outcomes
}
