% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{wins}
\alias{wins}
\title{Compute team wins}
\usage{
wins(tournament)
}
\arguments{
\item{tournament}{a tournament vector}
}
\value{
an integer vector containing the number of wins for each team
}
\description{
Use tournament results to compute number of wins for each team in the tournament
}
