% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_matrix}
\alias{scenario_matrix}
\title{Create Scenario Matrix}
\usage{
scenario_matrix(n_games = 15, max_games_remaining = 15)
}
\arguments{
\item{n_games}{number of games with undetermined outcome}

\item{max_games_remaining}{maximum number of games remaining (to avoid long computation time)}
}
\description{
Create a bit matrix of possible outcomes of unknown games.
Note: This gets large fast, so the input is capped at 15
}
\examples{
scenario_matrix(5)
}
