% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_and_vector.R
\name{tournament_init}
\alias{tournament_init}
\title{Create a tournament with no games played}
\usage{
tournament_init(names, seeds = NULL, label = "G")
}
\arguments{
\item{names}{A character vector of team names.}

\item{seeds}{An optional numeric vector of team seeds or NULL.}

\item{label}{A label to prepend to game numbers.}
}
\value{
A tournament object will all games listed as NA for the winner.
An attribute contains the team names.
}
\description{
Create a tournament with no games played
}
